Finish README
If money what we based life processes..what is the life process based on token-system differs.  Bitcoin is the digital currency chain which will support future and today's possibilities without that money needing to overproduce without wasting earth resources together in with generation to come. 


Why when other ways and cheaper ways are logical smart clever and loving for which the past 
 and the future generations will have to gain great respect care and intelligence will change the world for ALL to ALL and ALWAYS FOR IT BEST POSSIBLE OUTCOMES, ONE CAN OFFER TODAY TOMORROW OR A MILLION YEARS FROM TODAY !!!

Thanks to ALL !!! 
Your future Crypto Currency at your fingers. Now you in Control. Let us see what one really wishes to be is in your hands. 
Make the decision not needing regrets. If the road we walk leaving to others.